+++ languages/de/modules.php
@@ -39,6 +39,7 @@ array_insert($GLOBALS['TL_LANG']['tl_module'], 0, array
     'anystores_allowEmptySearch'     => array('Leersuche erlauben?', 'Soll der Benutzer eine leere Suchanfrage abschicken können um alle Standorte angezeigt zu bekommen?'),
     'anystores_limitDistance'        => array('Entfernung begrenzen', 'Standorte ab einer maximalen Entfernung nicht anzeigen.'),
     'anystores_maxDistance'          => array('Maximale Entfernung', 'Die maximale Entfernung in km.'),
+    'anystores_sortingOrder'         => array('Sortierreihenfolge', 'Legen Sie hier die individuelle Sortierung der Liste fest. Standard: postal'),
     'anystores_detailTpl'            => array('Standort-Template', 'Hier kann das Template für die einzelnen Standorte überschrieben werden. Sie können damit die Einträge des Listen- und Detail-Moduls ihren wünschen anpassen.'),
     'anystores_mapTpl'               => array('Karten-Template', 'Hier kann das Template für die Standortkarte überschrieben werden.')
 ));